requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
supabase==2.7.0
pydantic==2.6.0
python-dotenv==1.0.1
httpx==0.27.0